Move monitor resource into domain package
