+++ internal/provider/monitor/monitor_application_error_retries_test.go
@@ -1,4 +1,4 @@
-package provider
+package monitor
 
 import (
 	"context"
