+++ internal/provider/monitor/monitor_compare_test.go
@@ -1,4 +1,4 @@
-package provider
+package monitor
 
 import (
 	"testing"
