+++ internal/provider/monitor/monitor_ip_version_test.go
@@ -1,4 +1,4 @@
-package provider
+package monitor
 
 import (
 	"context"
