+++ internal/provider/monitor/monitor_plan_modify.go
@@ -1,4 +1,4 @@
-package provider
+package monitor
 
 import (
 	"context"
