+++ internal/provider/monitor/monitor_read_test.go
@@ -1,4 +1,4 @@
-package provider
+package monitor
 
 import (
 	"context"
