+++ internal/provider/monitor/monitor_region_data.go
@@ -1,4 +1,4 @@
-package provider
+package monitor
 
 import (
 	"context"
