+++ internal/provider/monitor/monitor_transform.go
@@ -1,4 +1,4 @@
-package provider
+package monitor
 
 import (
 	"bytes"
