+++ internal/provider/monitor/monitor_upgrade.go
@@ -1,4 +1,4 @@
-package provider
+package monitor
 
 import (
 	"context"
