+++ internal/provider/monitor/monitor_upgrader_helper.go
@@ -1,4 +1,4 @@
-package provider
+package monitor
 
 import (
 	"fmt"
