+++ internal/provider/monitor/upgrader_helpers_test.go
@@ -1,4 +1,4 @@
-package provider
+package monitor
 
 import (
 	"context"
