+++ internal/provider/provider.go
@@ -18,6 +18,7 @@ import (
 	"github.com/uptimerobot/terraform-provider-uptimerobot/internal/provider/integration"
 	"github.com/uptimerobot/terraform-provider-uptimerobot/internal/provider/iprange"
 	"github.com/uptimerobot/terraform-provider-uptimerobot/internal/provider/maintenancewindow"
+	"github.com/uptimerobot/terraform-provider-uptimerobot/internal/provider/monitor"
 	"github.com/uptimerobot/terraform-provider-uptimerobot/internal/provider/monitorgroup"
 	"github.com/uptimerobot/terraform-provider-uptimerobot/internal/provider/psp"
 	"github.com/uptimerobot/terraform-provider-uptimerobot/internal/provider/pspannouncement"
@@ -109,7 +110,7 @@ func (p *UptimeRobotProvider) Configure(ctx context.Context, req provider.Config
 
 func (p *UptimeRobotProvider) Resources(ctx context.Context) []func() resource.Resource {
 	return []func() resource.Resource{
-		NewMonitorResource,
+		monitor.NewResource,
 		monitorgroup.NewResource,
 		psp.NewResource,
 		pspannouncement.NewResource,
